requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.25.1
opencv-python==4.10.0.84
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.3
pillow==10.4.0
pip==24.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sympy==1.13.3
thop==0.1.1.post2209072238
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.1.29
urllib3==2.2.3
wheel==0.44.0
zipp==3.20.2
